feat: implement "store_non_parsable_chars" for consume_from_offset
When specified, this will allow the parsing function to attempt to skip
characters when parsing the stream so that it can parse grammars that
have problems in them.

This is important in an editor because as a user is editing the
document, it's going to change to be invalid syntax for brief points in
time and this mode is what will handle that.