fix: adjust where the parser prefers partial vs full matches
I'm running into this issue:
- repeat from 0 to infinity
  - any
    1. raw / regex that matches later in the string
    2. raw / regex

The problem is that #1 matches but with ::skipped chars before it. This then causes the parser to
get really confused because it can't parse the token stream from that point forward effectively and
lots of stuff gets marked as ::skipped

I think the foundational issue here is that in a repeat where min_repeats is 0, it should not be
good enough to get a partial match with stuff before it.