feat: fix major bug in % / MatchDelimeter matching of escaped parenthesis / square / curlys
Previously, there was a bug where going forwards, matching always would
match escaped symbol with an escaped symbol. However, when going
backwards, regular symbols would sometimes match with escaped symbols
because the system wouldn't look ahead for an escape backslash BEFORE
the symbol to make sure that it was not escaped (only checking to make
sure that a symbol WAS escaped when matching escaped to escaped was
occurring).

This change fixes this behavior so that now, escaped always match with
escaped, and unescaped always match with unescaped.