fix: address token changing issue with reparsing token literal
Previously, when the token tree was reparsed due to a token being
changed, the literal text being reparsed was not being changed. So that
meant that the literal text was constant and the new token parse's
results WEREN'T for the whole series of tokens that needed to be
reparsed, just for the first individual token.

To fix this, when attempting a reparse, update `working_new_text` to be
the whole contents of the new `working_token`.